Develop v1.1 (#5)
* IMGONNACRAZY!

* _develop_v1.1

* Swappers, Wolf Recruit & Role Plan

* _develop_v1.1_20240223

* _develop_v1.1_20240224

* _develop_v1.1_20240225

* UpdateV3

* Update Readme

* _develop_v1.1_20240228

* Translates

* _develop_v1.1_2020301

* About Neutrals

* develop_V1.1_20240306

* Supportv2024.3.5 && _develop_v1.1_20240209

* Preview_1

* merge with Town Of Host v5.1.5

* Release

+++ TONEX/Modules/SoundInterface/MusicDownloader.cs
@@ -32,7 +32,7 @@ public class MusicDownloader
     public static string downloadUrl_website2 = "";
     public static bool succeed = false;
 
-    public static void StartDownload(string sound, string url = "waitToSelect")
+    public static async Task StartDownload(string sound, string url = "waitToSelect")
     {
         succeed = false;
         if (!Directory.Exists(@"./TONEX_Data/Sounds"))
@@ -64,22 +64,48 @@ public class MusicDownloader
 
         try
         {
-
+           
             string filePath = DownloadFileTempPath;
-
-            HttpWebRequest request = (HttpWebRequest)WebRequest.Create(url);
-
-            using (HttpWebResponse response = (HttpWebResponse)request.GetResponse())
-            using (Stream streamResponse = response.GetResponseStream())
-            using (Stream streamFile = File.Create(filePath))
+            using var client = new HttpClientDownloadWithProgress(url, filePath);
+            client.ProgressChanged += OnDownloadProgressChanged;
+            await client.StartDownload();
+            Thread.Sleep(100);
+            
+            if (md5ForFiles.ContainsKey(sound))
             {
-                byte[] buffer = new byte[1024];
-                int bytesRead;
-
-                while ((bytesRead = streamResponse.Read(buffer, 0, buffer.Length)) > 0)
+                if (GetMD5HashFromFile(DownloadFileTempPath) != md5ForFiles[sound].ToLower())
                 {
-                    streamFile.Write(buffer, 0, bytesRead);
+                    Logger.Error($"Md5 Wrong in {url}", "DownloadSound");
+                    File.Delete(DownloadFileTempPath);
+                    if (url == downloadUrl_website)
+                    {
+                        url = downloadUrl_website2;
+                        goto retry;
+                    }
+                    else if (url == downloadUrl_website2)
+                    {
+                        url = downloadUrl_gitee;
+                        goto retry;
+                    }
+                    return;
                 }
+                Logger.Info($"Md5 Currect in {url}", "DownloadSound");
+            }
+            else
+            {
+                Logger.Error($"Md5 No Found in {url}", "DownloadSound");
+                File.Delete(DownloadFileTempPath);
+                if (url == downloadUrl_website)
+                {
+                    url = downloadUrl_website2;
+                    goto retry;
+                }
+                else if (url == downloadUrl_website2)
+                {
+                    url = downloadUrl_gitee;
+                    goto retry;
+                }
+                return;
             }
             succeed = true;
             Logger.Info($"Succeed in {url}", "DownloadSound");
@@ -87,7 +113,10 @@ public class MusicDownloader
         catch (Exception ex)
         {
             Logger.Error($"Failed to download\n{ex.Message}", "DownloadSound", false);
-            File.Delete(DownloadFileTempPath);
+            if (!string.IsNullOrEmpty(DownloadFileTempPath))
+            {
+                File.Delete(DownloadFileTempPath);
+            }
         }
         if (!succeed)
         {
@@ -103,12 +132,80 @@ public class MusicDownloader
             }
             
         }
-        
     }
     private static bool IsValidUrl(string url)
     {
         string pattern = @"^(https?|ftp)://[^\s/$.?#].[^\s]*$";
         return Regex.IsMatch(url, pattern);
     }
+    private static void OnDownloadProgressChanged(long? totalFileSize, long totalBytesDownloaded, double? progressPercentage)
+    {
+        string msg = $"{GetString("updateInProgress")}\n{totalFileSize / 1000}KB / {totalBytesDownloaded / 1000}KB  -  {(int)progressPercentage}%";
+        Logger.Info(msg, "DownloadDLL");
+    }
+    public static string GetMD5HashFromFile(string fileName)
+    {
+        try
+        {
+            using var md5 = MD5.Create();
+            using var stream = File.OpenRead(fileName);
+            var hash = md5.ComputeHash(stream);
+            return BitConverter.ToString(hash).Replace("-", "").ToLower();
+        }
+        catch (Exception ex)
+        {
+            Logger.Exception(ex, "GetMD5HashFromFile");
+            return "";
+        }
+    }
+    private static Dictionary<string, string> md5ForFiles = new()
+    {
+        {"AWP","A191F48B6689290ECD4568149C22A381" },
+        {"Bet","8B9B734E97998BE8872ADAE6B5D4343C"},
+        {"Bite","9AEFF327DE582FF926EF2187AE4DC338"},
+        {"Boom","4DF61F364E7EE087A983172021CEE00C"},
+        {"Clothe","394F4EC5823A7F8AD4ECEA6897D2928C"},
+        {"Congrats","65F53C4C1731B112CF5C38E6E1C74988"},
+        {"Curse","3548B2872E3630789FB664BE5137E3D3"},
+        {"Dove","C4FE25AF79505A866C8ECAB38761809F"},
+        {"Eat","4BBF93B90712722AC0DC3DD976634E78"},
+        {"ElementMaxi1","D99694C79BF36615939ED02FF05F339C"},
+        {"ElementMaxi2","F64D5C34ADE6637258DBC289BB47B59A"},
+        {"ElementMaxi3","D698A12A1801328739EE0B87777243AF"},
+        {"ElementSkill1","45204B00A499C52ACE852BFAE913076C"},
+        {"ElementSkill2","34892FB0B82C1A5A827AC955ED3147BF"},
+        {"ElementSkill3","BDF00B0AC80B4E6510619F2C9A5E2062"},
+        {"FlashBang","E4C9912E139F1DFFDFD95F0081472EBA"},
+        {"GongXiFaCai","7DED159AD441CA72DB98A442B037A3D7"},
+        
+        {"Gunfire","4A44EAF6B45B96B63BBC12A946DB517B"},
+        {"Gunload","27441FBFC8CC5A5F2945A8CE344A52B9"},
+        {"Join1","0DBC4FEDCD5C8D10A57EBB8E5C31189D"},
+        {"Join2","646B104360FD8DC2E20339291FC25BDE"},
+        {"Join3","E613F02735A761E720367AAED8F93AF9"},
+        {"Line","4DA0B66BD3E2C8D2D5984CB15F518378"},
+        {"MarioCoin","2698FB768F1E1045C1231B63C2639766"},
+        {"MarioJump","A485BCFEE7311EF3A7651F4B20E381CB"},
+        {"Onichian","13B71F389E21C2AF8E35996201843642"},
+        
+        {"Shapeshifter","B7119CC4E0E5B108B8735D734769AA5C"},
+        {"Shield","9EA3B450C5B53A4B952CB8418DF84539"},
+        
+        
+        {"Teleport","8D3DA143C59CD7C4060129C46BEB7A39"},
+        {"TheWorld","395010A373FAE0EC704BB4FE8FC5A57A"},
+        //“Ù¿÷
+        {"GongXiFaCaiLiuDeHua","DB200D93E613020D62645F4841DD55BD"},
+        {"RejoiceThisSEASONRespectThisWORLD","7AB4778744242E4CFA0468568308EA9B"},
+        {"SpringRejoicesinParallelUniverses","D92528104A82DBBFADB4FF251929BA5E"},
+        {"StarFallsWithDomeCrumbles_V4","EDCB0187C269BEF42A5E6EC25CFEADF7"},
+        {"TheDomeofTruth","183804914E3310B9F92B47392F503A9F"},
+        {"HopeStillExists_V2","8D5BA9AC283E156AB2C930F7B63A4A36"},
+        {"HeartGuidedbyLight","F1DED08A59936B8E1DB95067A69B006E"},
+        {"Thesorrowofpartingacrosslifetimes","D5071AF2BFE92465EACEB318E6FC3390"},
+        {"GuardianandDream","1AD45F97F70F05E19BCDF77C489932A6"},
+        {"AFamiliarPromise","A3672341F586B4D81EFBA6D4278CFEAE"},
+        {"DeterminationWithJustice","1AD45F97F70F05E19BCDF77C489932A6"},
 
+    };
 }
